Estimated meal logging

+++ routes/api.php
@@ -9,7 +9,6 @@ use App\Http\Controllers\UserController;
 use App\Http\Controllers\UserProfileController;
 use Illuminate\Support\Facades\Route;
 
-
 Route::prefix('v1')->group(function () {
 
     Route::prefix('auth')->group(function () {
@@ -43,16 +42,18 @@ Route::prefix('v1')->group(function () {
             Route::delete('/avatar', [UserController::class, 'destroyAvatar']);
             Route::post('/cover-image', [UserController::class, 'storeCoverImage']);
             Route::delete('/cover-image', [UserController::class, 'destroyCoverImage']);
-            Route::post('/log/{meal}', [DailyLogingController::class, 'logMeal']);
             Route::post('/log', [DailyLogingController::class, 'logCustomMeal']);
-            Route::delete('/log/{log}', [DailyLogingController::class, 'removeDailyLog']);
+            Route::post('/log/estimate', [DailyLogingController::class, 'LogEstimatedMeal']);
+            Route::post('/log/{meal}', [DailyLogingController::class, 'logMeal']);
+            Route::delete('/log/{log}', [DailyLogingController::class, 'removeDailyLog'])->middleware('ensure.owns:log');
         });
 
         Route::prefix('meals')->group(function () {
             Route::post('/', [MealController::class, 'store']);
-            Route::post('/{meal}/confirm', [MealController::class, 'confirm']);
-            Route::post('/{meal}/discard', [MealController::class, 'discard']);
+            Route::post('/{meal}/confirm', [MealController::class, 'confirm'])->middleware('ensure.owns:meal,user_profile_id');
+            Route::post('/{meal}/discard', [MealController::class, 'discard'])->middleware('ensure.owns:meal,user_profile_id');
         });
+
 
         Route::prefix('ingredients')->group(function () {
             Route::post('/search', [IngredientController::class, 'search']);
